move test helpers to common lib
