+++ ydb/core/tx/datashard/datashard_ut_snapshot_isolation.cpp
@@ -1,15 +1,15 @@
 #include "datashard_ut_common_kqp.h"
-#include <ydb/core/tx/data_events/payload_helper.h>
 #include <ydb/core/tx/datashard/ut_common/datashard_ut_common.h>
+#include <ydb/core/tx/datashard/ut_common/datashard_ut_common_tx.h>
 #include <ydb/core/base/tablet_pipecache.h>
 #include <ydb/core/testlib/actors/block_events.h>
 #include <ydb/core/protos/query_stats.pb.h>
-#include <ydb/core/tx/long_tx_service/public/lock_handle.h>
 
 namespace NKikimr {
 
 using namespace NKikimr::NDataShard;
 using namespace NKikimr::NDataShard::NKqpHelpers;
+using namespace NKikimr::NDataShard::NTxHelpers;
 using namespace NSchemeShard;
 using namespace Tests;
 
@@ -27,437 +27,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         return {runtime, server, sender};
     }
 
-    ui64 AllocateTxId(TTestActorRuntime& runtime, const TActorId& sender) {
-        ui32 nodeId = sender.NodeId();
-        ui32 nodeIndex = nodeId - runtime.GetNodeId(0);
-        runtime.Send(new IEventHandle(MakeTxProxyID(), sender, new TEvTxUserProxy::TEvAllocateTxId), nodeIndex, true);
-        auto ev = runtime.GrabEdgeEventRethrow<TEvTxUserProxy::TEvAllocateTxIdResult>(sender);
-        auto* msg = ev->Get();
-        return msg->TxId;
-    }
-
-    TRowVersion AcquireSnapshot(TTestActorRuntime& runtime, const TActorId& sender, const TString& databaseName = "/Root") {
-        ui32 nodeId = sender.NodeId();
-        ui32 nodeIndex = nodeId - runtime.GetNodeId(0);
-        auto* req = new NLongTxService::TEvLongTxService::TEvAcquireReadSnapshot(databaseName);
-        runtime.Send(new IEventHandle(NLongTxService::MakeLongTxServiceID(nodeId), sender, req), nodeIndex, true);
-        auto ev = runtime.GrabEdgeEventRethrow<TEvLongTxService::TEvAcquireReadSnapshotResult>(sender);
-        auto* msg = ev->Get();
-        UNIT_ASSERT_VALUES_EQUAL(msg->Status, Ydb::StatusIds::SUCCESS);
-        return msg->Snapshot;
-    }
-
-    NLongTxService::TLockHandle MakeLockHandle(TTestActorRuntime& runtime, ui64 lockId, ui32 nodeIndex = 0) {
-        return NLongTxService::TLockHandle(lockId, runtime.GetActorSystem(nodeIndex));
-    }
-
-    TString FormatReadResult(const TEvDataShard::TEvReadResult* msg) {
-        TStringBuilder sb;
-        if (msg->Record.GetStatus().GetCode() == Ydb::StatusIds::SUCCESS) {
-            size_t count = msg->GetRowsCount();
-            for (size_t i = 0; i < count; ++i) {
-                auto cells = msg->GetCells(i);
-                for (size_t j = 0; j < cells.size(); ++j) {
-                    if (j != 0) {
-                        sb << ", ";
-                    }
-                    sb << cells[j].AsValue<i32>();
-                }
-                sb << "\n";
-            }
-        } else {
-            sb << "ERROR: " << msg->Record.GetStatus().GetCode();
-        }
-        return sb;
-    }
-
-    struct TKeyValue {
-        i32 Key;
-        i32 Value;
-    };
-
-    struct TOperation {
-        NKikimrDataEvents::TEvWrite::TOperation::EOperationType Type;
-        TVector<TKeyValue> Rows;
-
-        void ApplyTo(const TTableId& tableId, NEvents::TDataEvents::TEvWrite* req) {
-            TVector<TCell> cells;
-            cells.reserve(Rows.size() * 2);
-            for (const auto& row : Rows) {
-                cells.push_back(TCell::Make(row.Key));
-                cells.push_back(TCell::Make(row.Value));
-            }
-
-            TSerializedCellMatrix matrix(cells, Rows.size(), 2);
-            TString blobData = matrix.ReleaseBuffer();
-
-            ui64 payloadIndex = NKikimr::NEvWrite::TPayloadWriter<NKikimr::NEvents::TDataEvents::TEvWrite>(*req).AddDataToPayload(std::move(blobData));
-            req->AddOperation(Type, tableId, { 1, 2 }, payloadIndex, NKikimrDataEvents::FORMAT_CELLVEC);
-        }
-
-        static TOperation Upsert(i32 key, i32 value) {
-            return Upsert({ { key, value } });
-        }
-
-        static TOperation Upsert(TVector<TKeyValue> rows) {
-            return TOperation{
-                NKikimrDataEvents::TEvWrite::TOperation::OPERATION_UPSERT,
-                std::move(rows),
-            };
-        }
-
-        static TOperation Insert(i32 key, i32 value) {
-            return Insert({ { key, value } } );
-        }
-
-        static TOperation Insert(TVector<TKeyValue> rows) {
-            return TOperation{
-                NKikimrDataEvents::TEvWrite::TOperation::OPERATION_INSERT,
-                std::move(rows),
-            };
-        }
-    };
-
-    class TTransactionState {
-    public:
-        TTransactionState(TTestActorRuntime& runtime)
-            : Runtime(runtime)
-        {
-            Sender = Runtime.AllocateEdgeActor();
-            LockTxId = AllocateTxId(Runtime, Sender);
-            LockNodeId = Sender.NodeId();
-            LockHandle = MakeLockHandle(Runtime, LockTxId, LockNodeId - Runtime.GetNodeId(0));
-        }
-
-        struct TReadPromise {
-            TTransactionState& State;
-            TActorId Sender;
-            TActorId LastSender = TActorId();
-            ui64 LastSeqNo = 0;
-
-            void SendAck(ui64 maxRows = Max<ui64>(), ui64 maxBytes = Max<ui64>()) {
-                auto* msg = new TEvDataShard::TEvReadAck();
-                msg->Record.SetReadId(0);
-                msg->Record.SetSeqNo(LastSeqNo);
-                msg->Record.SetMaxRows(maxRows);
-                msg->Record.SetMaxBytes(maxBytes);
-                ui32 nodeIndex = Sender.NodeId() - State.Runtime.GetNodeId(0);
-                State.Runtime.Send(new IEventHandle(LastSender, Sender, msg), nodeIndex, true);
-            }
-
-            std::unique_ptr<TEvDataShard::TEvReadResult> NextResult(TDuration simTimeout = TDuration::Max()) {
-                auto ev = State.Runtime.GrabEdgeEventRethrow<TEvDataShard::TEvReadResult>(Sender, simTimeout);
-                if (!ev) {
-                    return nullptr;
-                }
-                LastSender = ev->Sender;
-                std::unique_ptr<TEvDataShard::TEvReadResult> msg(ev->Release().Release());
-                LastSeqNo = msg->Record.GetSeqNo();
-                for (const auto& lock : msg->Record.GetTxLocks()) {
-                    State.Locks.push_back(lock);
-                }
-                for (const auto& lock : msg->Record.GetBrokenTxLocks()) {
-                    State.Locks.push_back(lock);
-                }
-                return msg;
-            }
-
-            TString NextString(TDuration simTimeout = TDuration::Max()) {
-                auto msg = NextResult(simTimeout);
-                if (!msg) {
-                    return "<timeout>";
-                }
-                auto status = msg->Record.GetStatus().GetCode();
-                if (status != Ydb::StatusIds::SUCCESS) {
-                    return TStringBuilder() << "ERROR: " << status;
-                }
-                TString res = FormatReadResult(msg.get());
-                if (msg->Record.GetFinished()) {
-                    res += "<end>";
-                }
-                return res;
-            }
-
-            TString AllString() {
-                TString res;
-                for (;;) {
-                    auto msg = NextResult();
-                    auto status = msg->Record.GetStatus().GetCode();
-                    if (status != Ydb::StatusIds::SUCCESS) {
-                        res += TStringBuilder() << "ERROR: " << status;
-                        break;
-                    }
-                    res += FormatReadResult(msg.get());
-                    if (msg->Record.GetFinished()) {
-                        break;
-                    }
-                    SendAck();
-                }
-                return res;
-            }
-        };
-
-        TReadPromise SendReadKey(const TTableId& tableId, ui64 shardId, i32 key) {
-            if (!Snapshot) {
-                Snapshot = AcquireSnapshot(Runtime, Sender);
-            }
-
-            TActorId sender = Runtime.AllocateEdgeActor();
-            ui32 nodeIndex = sender.NodeId() - Runtime.GetNodeId(0);
-
-            auto* req = new TEvDataShard::TEvRead();
-            req->Record.SetReadId(0);
-            req->Record.MutableTableId()->SetOwnerId(tableId.PathId.OwnerId);
-            req->Record.MutableTableId()->SetTableId(tableId.PathId.LocalPathId);
-            req->Record.MutableTableId()->SetSchemaVersion(tableId.SchemaVersion);
-            req->Record.MutableSnapshot()->SetStep(Snapshot->Step);
-            req->Record.MutableSnapshot()->SetTxId(Snapshot->TxId);
-            req->Record.SetLockTxId(LockTxId);
-            req->Record.SetLockNodeId(LockNodeId);
-            req->Record.SetLockMode(NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
-            req->Record.AddColumns(1);
-            req->Record.AddColumns(2);
-            req->Record.SetResultFormat(NKikimrDataEvents::FORMAT_CELLVEC);
-
-            TVector<TCell> cells;
-            cells.emplace_back(TCell::Make(key));
-            req->Keys.emplace_back(cells);
-
-            Runtime.SendToPipe(shardId, sender, req, nodeIndex);
-            return { *this, sender };
-        }
-
-        TString ReadKey(const TTableId& tableId, ui64 shardId, i32 key) {
-            auto promise = SendReadKey(tableId, shardId, key);
-            return promise.AllString();
-        }
-
-        TReadPromise SendReadRange(const TTableId& tableId, ui64 shardId, i32 minKey, i32 maxKey, ui64 maxRowsQuota = Max<ui64>()) {
-            if (!Snapshot) {
-                Snapshot = AcquireSnapshot(Runtime, Sender);
-            }
-
-            TActorId sender = Runtime.AllocateEdgeActor();
-            ui32 nodeIndex = sender.NodeId() - Runtime.GetNodeId(0);
-
-            auto* req = new TEvDataShard::TEvRead();
-            req->Record.SetReadId(0);
-            req->Record.MutableTableId()->SetOwnerId(tableId.PathId.OwnerId);
-            req->Record.MutableTableId()->SetTableId(tableId.PathId.LocalPathId);
-            req->Record.MutableTableId()->SetSchemaVersion(tableId.SchemaVersion);
-            req->Record.MutableSnapshot()->SetStep(Snapshot->Step);
-            req->Record.MutableSnapshot()->SetTxId(Snapshot->TxId);
-            req->Record.SetLockTxId(LockTxId);
-            req->Record.SetLockNodeId(LockNodeId);
-            req->Record.SetLockMode(NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
-            req->Record.AddColumns(1);
-            req->Record.AddColumns(2);
-            req->Record.SetResultFormat(NKikimrDataEvents::FORMAT_CELLVEC);
-            req->Record.SetMaxRowsInResult(1);
-            req->Record.SetMaxRows(maxRowsQuota);
-
-            TVector<TCell> minCells, maxCells;
-            minCells.emplace_back(TCell::Make(minKey));
-            maxCells.emplace_back(TCell::Make(maxKey));
-            req->Ranges.emplace_back(minCells, true, maxCells, true);
-
-            Runtime.SendToPipe(shardId, sender, req, nodeIndex);
-            return { *this, sender };
-        }
-
-        TString ReadRange(const TTableId& tableId, ui64 shardId, i32 minKey, i32 maxKey) {
-            auto promise = SendReadRange(tableId, shardId, minKey, maxKey);
-            return promise.AllString();
-        }
-
-        bool HasLockConflicts(ui64 shardId) const {
-            const NKikimrDataEvents::TLock* prev = nullptr;
-            for (auto it = Locks.begin(); it != Locks.end(); ++it) {
-                if (it->GetDataShard() == shardId) {
-                    if (prev) {
-                        if (prev->GetGeneration() != it->GetGeneration()) {
-                            return true;
-                        }
-                        if (prev->GetCounter() != it->GetCounter()) {
-                            return true;
-                        }
-                    }
-                    prev = &*it;
-                }
-            }
-            return false;
-        }
-
-        const NKikimrDataEvents::TLock* FindLastLock(ui64 shardId) const {
-            for (auto it = Locks.rbegin(); it != Locks.rend(); ++it) {
-                if (it->GetDataShard() == shardId) {
-                    return &*it;
-                }
-            }
-            return nullptr;
-        }
-
-        struct TWritePromise {
-            TTransactionState& State;
-            TActorId Sender;
-
-            std::unique_ptr<NEvents::TDataEvents::TEvWriteResult> NextResult(TDuration simTimeout = TDuration::Max()) {
-                auto ev = State.Runtime.GrabEdgeEventRethrow<NEvents::TDataEvents::TEvWriteResult>(Sender, simTimeout);
-                if (!ev) {
-                    return nullptr;
-                }
-                std::unique_ptr<NEvents::TDataEvents::TEvWriteResult> msg(ev->Release().Release());
-                for (const auto& lock : msg->Record.GetTxLocks()) {
-                    State.Locks.push_back(lock);
-                }
-                return msg;
-            }
-
-            TString NextString(TDuration simTimeout = TDuration::Max()) {
-                auto msg = NextResult(simTimeout);
-                if (!msg) {
-                    return "<timeout>";
-                }
-                auto status = msg->Record.GetStatus();
-                if (status != NKikimrDataEvents::TEvWriteResult::STATUS_COMPLETED) {
-                    return TStringBuilder() << "ERROR: " << status;
-                }
-                return "OK";
-            }
-        };
-
-        template<class... TOps>
-        TWritePromise SendWrite(const TTableId& tableId, ui64 shardId, TOps&&... ops) {
-            auto sender = Runtime.AllocateEdgeActor();
-            ui32 nodeIndex = sender.NodeId() - Runtime.GetNodeId(0);
-
-            auto* req = new NEvents::TDataEvents::TEvWrite(0, NKikimrDataEvents::TEvWrite::MODE_IMMEDIATE);
-            req->Record.SetLockTxId(LockTxId);
-            req->Record.SetLockNodeId(LockNodeId);
-            req->Record.SetLockMode(NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
-            if (Snapshot) {
-                req->Record.MutableMvccSnapshot()->SetStep(Snapshot->Step);
-                req->Record.MutableMvccSnapshot()->SetTxId(Snapshot->TxId);
-            }
-
-            (..., ops.ApplyTo(tableId, req));
-
-            Runtime.SendToPipe(shardId, sender, req, nodeIndex);
-            return { *this, sender };
-        }
-
-        template<class... TOps>
-        TString Write(const TTableId& tableId, ui64 shardId, TOps&&... ops) {
-            auto promise = SendWrite(tableId, shardId, std::forward<TOps>(ops)...);
-            return promise.NextString();
-        }
-
-        template<class... TOps>
-        TWritePromise SendWriteCommit(const TTableId& tableId, ui64 shardId, TOps&&... ops) {
-            auto sender = Runtime.AllocateEdgeActor();
-            ui32 nodeIndex = sender.NodeId() - Runtime.GetNodeId(0);
-
-            auto* req = new NEvents::TDataEvents::TEvWrite(0, NKikimrDataEvents::TEvWrite::MODE_IMMEDIATE);
-            req->Record.SetLockMode(NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
-            if (Snapshot) {
-                req->Record.MutableMvccSnapshot()->SetStep(Snapshot->Step);
-                req->Record.MutableMvccSnapshot()->SetTxId(Snapshot->TxId);
-            }
-
-            // Try to find the last known lock state
-            if (const auto* pLock = FindLastLock(shardId)) {
-                req->Record.MutableLocks()->SetOp(NKikimrDataEvents::TKqpLocks::Commit);
-                req->Record.MutableLocks()->AddSendingShards(shardId);
-                req->Record.MutableLocks()->AddReceivingShards(shardId);
-                *req->Record.MutableLocks()->AddLocks() = *pLock;
-            }
-
-            (..., ops.ApplyTo(tableId, req));
-
-            Runtime.SendToPipe(shardId, sender, req, nodeIndex);
-            return { *this, sender };
-        }
-
-        template<class... TOps>
-        TString WriteCommit(const TTableId& tableId, ui64 shardId, TOps&&... ops) {
-            auto promise = SendWriteCommit(tableId, shardId, std::forward<TOps>(ops)...);
-            return promise.NextString();
-        }
-
-        void InitCommit(std::vector<ui64> participants) {
-            CommitTxId = AllocateTxId(Runtime, Sender);
-            Participants = std::move(participants);
-        }
-
-        template<class... TOps>
-        TWritePromise SendPrepareCommit(const TTableId& tableId, ui64 shardId, TOps&&... ops) {
-            auto sender = Runtime.AllocateEdgeActor();
-            ui32 nodeIndex = sender.NodeId() - Runtime.GetNodeId(0);
-
-            auto* req = new NEvents::TDataEvents::TEvWrite(CommitTxId, NKikimrDataEvents::TEvWrite::MODE_VOLATILE_PREPARE);
-            req->Record.SetLockMode(NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
-            if (Snapshot) {
-                req->Record.MutableMvccSnapshot()->SetStep(Snapshot->Step);
-                req->Record.MutableMvccSnapshot()->SetTxId(Snapshot->TxId);
-            }
-
-            req->Record.MutableLocks()->SetOp(NKikimrDataEvents::TKqpLocks::Commit);
-            for (ui64 participant : Participants) {
-                req->Record.MutableLocks()->AddSendingShards(participant);
-                req->Record.MutableLocks()->AddReceivingShards(participant);
-            }
-            if (const auto* pLock = FindLastLock(shardId)) {
-                *req->Record.MutableLocks()->AddLocks() = *pLock;
-            }
-
-            (..., ops.ApplyTo(tableId, req));
-
-            Runtime.SendToPipe(shardId, sender, req, nodeIndex);
-            return { *this, sender };
-        }
-
-        template<class... TOps>
-        TWritePromise PrepareCommit(const TTableId& tableId, ui64 shardId, TOps&&... ops) {
-            auto promise = SendPrepareCommit(tableId, shardId, std::forward<TOps>(ops)...);
-            auto msg = promise.NextResult();
-            UNIT_ASSERT_VALUES_EQUAL(msg->Record.GetStatus(), NKikimrDataEvents::TEvWriteResult::STATUS_PREPARED);
-            MinStep = Max(MinStep, msg->Record.GetMinStep());
-            MaxStep = Min(MaxStep, msg->Record.GetMaxStep());
-            for (ui64 coordinator : msg->Record.GetDomainCoordinators()) {
-                Coordinator = coordinator;
-            }
-            return promise;
-        }
-
-        void SendPlan() {
-            UNIT_ASSERT(Coordinator != 0);
-            UNIT_ASSERT(MinStep <= MaxStep);
-            UNIT_ASSERT(!Participants.empty());
-            SendProposeToCoordinator(
-                Runtime, Sender, Participants, {
-                    .TxId = CommitTxId,
-                    .Coordinator = Coordinator,
-                    .MinStep = MinStep,
-                    .MaxStep = MaxStep,
-                    .Volatile = true,
-                });
-        }
-
-    public:
-        TTestActorRuntime& Runtime;
-        TActorId Sender;
-        ui64 LockTxId = 0;
-        ui32 LockNodeId = 0;
-        NLongTxService::TLockHandle LockHandle;
-        std::optional<TRowVersion> Snapshot;
-        ui64 Counter = 0;
-        std::vector<NKikimrDataEvents::TLock> Locks;
-        std::vector<ui64> Participants;
-        ui64 CommitTxId = 0;
-        ui64 MinStep = 0;
-        ui64 MaxStep = Max<ui64>();
-        ui64 Coordinator = 0;
-    };
+    using TOperation = NKikimr::NDataShard::NTxHelpers::TWriteOperation;
 
     Y_UNIT_TEST(ReadWriteNoLocksNoConflict) {
         TPortManager pm;
@@ -487,7 +57,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 1u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -542,7 +112,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 1u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -590,7 +160,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 1u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -646,7 +216,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 1u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -654,7 +224,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             "1, 1001\n");
         UNIT_ASSERT_VALUES_EQUAL(tx1.Locks.size(), 0u);
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Make an uncommitted write in tx2 to key 2
         UNIT_ASSERT_VALUES_EQUAL(
@@ -717,7 +287,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 1u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -781,7 +351,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 1u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -838,7 +408,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 1u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -896,7 +466,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -910,7 +480,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             return !(msg->Record.GetFlags() & NKikimrTx::TEvReadSet::FLAG_EXPECT_READSET);
         });
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
         tx2.InitCommit(shards);
         auto write1 = tx2.PrepareCommit(tableId, shards.at(0), TOperation::Upsert(2, 2001));
         auto write2 = tx2.PrepareCommit(tableId, shards.at(1), TOperation::Upsert(11, 2002));
@@ -972,7 +542,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -986,7 +556,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             return !(msg->Record.GetFlags() & NKikimrTx::TEvReadSet::FLAG_EXPECT_READSET);
         });
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
         tx2.InitCommit(shards);
         auto write1 = tx2.PrepareCommit(tableId, shards.at(0), TOperation::Upsert(2, 2001));
         auto write2 = tx2.PrepareCommit(tableId, shards.at(1), TOperation::Upsert(11, 2002));
@@ -1062,7 +632,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1076,7 +646,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             return !(msg->Record.GetFlags() & NKikimrTx::TEvReadSet::FLAG_EXPECT_READSET);
         });
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
         tx2.InitCommit(shards);
         auto write1 = tx2.PrepareCommit(tableId, shards.at(0), TOperation::Upsert(2, 2001));
         auto write2 = tx2.PrepareCommit(tableId, shards.at(1), TOperation::Upsert(11, 2002));
@@ -1139,7 +709,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1153,7 +723,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             return !(msg->Record.GetFlags() & NKikimrTx::TEvReadSet::FLAG_EXPECT_READSET);
         });
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
         tx2.InitCommit(shards);
         auto write1 = tx2.PrepareCommit(tableId, shards.at(0), TOperation::Upsert(2, 2001));
         auto write2 = tx2.PrepareCommit(tableId, shards.at(1), TOperation::Upsert(11, 2002));
@@ -1225,7 +795,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             return !(msg->Record.GetFlags() & NKikimrTx::TEvReadSet::FLAG_EXPECT_READSET);
         });
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
         tx1.InitCommit(shards);
         auto write1 = tx1.PrepareCommit(tableId, shards.at(0), TOperation::Upsert(2, 2001));
         auto write2 = tx1.PrepareCommit(tableId, shards.at(1), TOperation::Upsert(11, 2002));
@@ -1234,7 +804,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         runtime.WaitFor("blocked readsets", [&]{ return blockedReadSets.size() >= 2; });
         blockedReadSets.Stop();
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1300,7 +870,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             return !(msg->Record.GetFlags() & NKikimrTx::TEvReadSet::FLAG_EXPECT_READSET);
         });
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
         tx1.InitCommit(shards);
         auto write1 = tx1.PrepareCommit(tableId, shards.at(0), TOperation::Upsert(2, 2001));
         auto write2 = tx1.PrepareCommit(tableId, shards.at(1), TOperation::Upsert(11, 2002));
@@ -1309,7 +879,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         runtime.WaitFor("blocked readsets", [&]{ return blockedReadSets.size() >= 2; });
         blockedReadSets.Stop();
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1372,7 +942,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
             return !(msg->Record.GetFlags() & NKikimrTx::TEvReadSet::FLAG_EXPECT_READSET);
         });
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
         tx1.InitCommit(shards);
         auto write1 = tx1.PrepareCommit(tableId, shards.at(0), TOperation::Upsert(2, 2001));
         auto write2 = tx1.PrepareCommit(tableId, shards.at(1), TOperation::Upsert(11, 2002));
@@ -1381,7 +951,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         runtime.WaitFor("blocked readsets", [&]{ return blockedReadSets.size() >= 2; });
         blockedReadSets.Stop();
 
-        TTransactionState tx2(runtime);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1440,7 +1010,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1488,7 +1058,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Start transaction by reading key 1, this should not acquire locks
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1536,7 +1106,7 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Read range, we should observe initial values
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1604,8 +1174,8 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
-        TTransactionState tx2(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Make an uncommitted write to key 2 in tx1
         UNIT_ASSERT_VALUES_EQUAL(
@@ -1666,8 +1236,8 @@ Y_UNIT_TEST_SUITE(DataShardSnapshotIsolation) {
         const auto shards = GetTableShards(server, sender, "/Root/table");
         UNIT_ASSERT_VALUES_EQUAL(shards.size(), 2u);
 
-        TTransactionState tx1(runtime);
-        TTransactionState tx2(runtime);
+        TTransactionState tx1(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
+        TTransactionState tx2(runtime, NKikimrDataEvents::OPTIMISTIC_SNAPSHOT_ISOLATION);
 
         // Make an uncommitted write to key 2 in tx1
         UNIT_ASSERT_VALUES_EQUAL(
