+++ ydb/core/tx/datashard/ut_common/datashard_ut_common.cpp
@@ -2934,4 +2934,24 @@ ui64 AsyncTruncateTable(
     return RunSchemeTx(*server->GetRuntime(), std::move(request), sender);
 }
 
+TString FormatReadResult(const TEvDataShard::TEvReadResult* msg) {
+    TStringBuilder sb;
+    if (msg->Record.GetStatus().GetCode() == Ydb::StatusIds::SUCCESS) {
+        size_t count = msg->GetRowsCount();
+        for (size_t i = 0; i < count; ++i) {
+            auto cells = msg->GetCells(i);
+            for (size_t j = 0; j < cells.size(); ++j) {
+                if (j != 0) {
+                    sb << ", ";
+                }
+                sb << cells[j].AsValue<i32>();
+            }
+            sb << "\n";
+        }
+    } else {
+        sb << "ERROR: " << msg->Record.GetStatus().GetCode();
+    }
+    return sb;
+}
+
 }
