+++ ydb/core/tx/datashard/ut_common/ya.make
@@ -11,6 +11,8 @@ YQL_LAST_ABI_VERSION()
 SRCS(
     datashard_ut_common.cpp
     datashard_ut_common.h
+    datashard_ut_common_tx.cpp
+    datashard_ut_common_tx.h
 )
 
 END()
